Implemented Base Maploader
Utilises the Gleed2d editor to load in maps via XML.

4 layers.
Background
Objects
Foreground
Collision

Background, object and foreground layers are just sprites that are
stored in vectors in the level class and then rendered in the correct
order. The collision layer is a vector a rectangles that define where
the collisionbounds of the level are. In Gleed2D this is drawn by using
the rectangle primitive.

Images are stored in an std::map so that duplicate images don't get
loaded in. (In SFML, images are loaded into memory and then assigned to
a sprite, there is no point loading in "GrassTile.png" 10 times to
render 10 grasstile sprites, you just reuse the image. It's handled
automatically here so no need to worrk.