Fixed the damn timestep stutter
The accumulator approach was dumb,(not really, it's pretty smart when
implemented properly but it wasn't in this case.) I took it out because
it was causing a stutter every 60 frames or so(caused by 2 simulation
steps being done at once.) We're back to a more familiar, multiply
deltaTime in based approach. I was confused before by the difference in
implementation between velocity and acceleration, all sorted now.