Implemented destructible objects, + refactoring
Implemented destructible objects, can be created in the map editor and
the game will handle it, still need to write instructions though, as you
need to have another image with -Destructed as a tag on it. Attack is
bound to K for the moment.

Fixed a bug where when you would drag the window, it would cause a
humungous delta time when you let go, cause windows pauses the
simulation when you're dragging a window. This meant that objects would
jump huge distances to try and make up for the framerate discrepency,
leading to collision-tunneling.

Also refactored the code where the player config file is loaded to use
one generic function to load in a numeric value rather than repreating
the same 5 lines of code 10 times over.