Implemented double jump
Player can now double jump. Also fixed the sprite aliasing issue by
changing the size of mainCharacter.png to be even numbers. Don't know
why it works, but it does.

When adding a new ability, make sure you update the ToggleAbility()
function as well as the enum so we can easily add/remove player ability
as the level goes along.