Gleed Fixer + New Level Layers
Wrote an application that should fix the content root folder issues
automatically for the Gleed editor. So the designers dont have to much
about in XML. Also, the level now loads in 3 background layers, far, mid
and near, so we can have parralax effects, even though that's not
implemented yet.