Falling Platforms + Parralax
Added support for falling platforms, addable from the map editor, with
DEATHZONES to destroy them when they fall, which will also be useful for
player respawn. Of course the fall values for this are configurable from
levelConfig.xml

Parralax between the far, mid and near background layers is also in now,
adding a lot of depth. Again, the values for this are configurable in
levelConfig.xml