Moving Platforms, Foreground Layer, More Parralax.
Altered the falling platforms to be a more generic special platform.
Gave them the ability to move with the player retaining velocity on them
so it functionally the player can ride them (This was bloody hard.)
The path of the moving platform can be drawn using the Gleed path tool,
and can be set to either loop or ping-pong the platforms. The speed of
the platforms, what path they use, as well as whether or not they
fall/move can also be set in the editor using custom properties.