Fixed the worse bug ever concerning menu tweens
Seriously, wasn't setting a boolean in the menuButton class, but the
thing is, it works fine in debug/release in VS with the debugger
attatched, and fine in debug without the debugger attactced(ie,
launching from .exe.) But in release without the debugger, it SOMETIMES
dosent tween in SOME menu elements. Had to attatch a debugger after
launching the game and do some fucking tedious debugging, all because i
forgot this->shouldTweenIn = shouldTweenIn.