All levels in and chained, chars working
All 7 levels from prologue to ending are in with their own audio config
files, and look to be working fine.
Updated the character configs to make them work with the new features
Added a guard from NaNs on player move to try and fix "The Bug." No clue
if it actually has though, more testing definately required.